seperate the 2 features to different files, fix PR comments
