+++ src/ds/features/auto-exposure-limit-feature.h
@@ -7,38 +7,20 @@
 #include <src/feature-interface.h>
 
 #include <memory>
+#include <librealsense2/hpp/rs_options.hpp>
 
 namespace librealsense {
 
 class synthetic_sensor;
 class hw_monitor;
+class limits_option;
 
-class range_limit_feature : public feature_interface
-{
-public:
-    static const feature_id ID;
-
-    range_limit_feature() = default;
-
-    feature_id get_id() const override;
-};
-
-class auto_exposure_limit_feature : public range_limit_feature
+class auto_exposure_limit_feature : public feature_interface
 {
 public:
     static const feature_id ID;
 
     auto_exposure_limit_feature( synthetic_sensor & depth_sensor, std::shared_ptr< hw_monitor > hw_monitor );
-
-    feature_id get_id() const override;
-};
-
-class gain_limit_feature : public range_limit_feature
-{
-public:
-    static const feature_id ID;
-
-    gain_limit_feature( synthetic_sensor & depth_sensor, std::shared_ptr< hw_monitor > hw_monitor );
 
     feature_id get_id() const override;
 };
